Corrects a bug in ukbb2csv's death censoring. When civic.Date was replaced with null.Date, the approach of manually setting the entry.died.Time no longer triggered the entry.died.Valid = true setting.
Now we entry.died.Scan() the date of death in order to also set
entry.died.Valid. Therefore, the phenotype and death censor dates
are now correctly computed. Before, the death_age would be correct,
but the death_censor_age and phenotype_censor_age would not be
influenced by the death event.